Added the configuration for the segment
