+++ eslint.config.mjs
@@ -1,9 +1,9 @@
 // For more info, see https://github.com/storybookjs/eslint-plugin-storybook#configuration-flat-config-format
-import storybook from "eslint-plugin-storybook";
+import storybook from 'eslint-plugin-storybook';
+import { dirname } from 'path';
+import { fileURLToPath } from 'url';
 
-import { dirname } from "path";
-import { fileURLToPath } from "url";
-import { FlatCompat } from "@eslint/eslintrc";
+import { FlatCompat } from '@eslint/eslintrc';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -15,20 +15,20 @@ const compat = new FlatCompat({
 const eslintConfig = [
   {
     ignores: [
-      ".next/**/*",
-      "node_modules/**/*",
-      ".storybook/**/*",
-      "coverage/**/*",
-      "dist/**/*",
-      "build/**/*",
-      "src/**/*.test.*"
-    ]
+      '.next/**/*',
+      'node_modules/**/*',
+      '.storybook/**/*',
+      'coverage/**/*',
+      'dist/**/*',
+      'build/**/*',
+      'src/**/*.test.*',
+    ],
   },
-  ...compat.extends("next/core-web-vitals", "next/typescript"),
-  ...storybook.configs["flat/recommended"],
+  ...compat.extends('next/core-web-vitals', 'next/typescript'),
+  ...storybook.configs['flat/recommended'],
   {
-    files: ["src/**/*.{js,jsx,ts,tsx}"],
-    
+    files: ['src/**/*.{js,jsx,ts,tsx}'],
+
     rules: {
       "@typescript-eslint/no-explicit-any": "error",
       "@typescript-eslint/no-unused-vars": ["error", {
