+++ package.json
@@ -27,6 +27,7 @@
     "@mui/x-date-pickers": "^8.12.0",
     "@mui/x-date-pickers-pro": "^8.11.2",
     "@reduxjs/toolkit": "^2.8.2",
+    "@segment/analytics-next": "^1.81.1",
     "@types/qrcode": "^1.5.5",
     "dayjs": "^1.11.18",
     "framer-motion": "^12.23.12",
