+++ src/app/(main)/contact-us/type.ts
@@ -1,6 +1,5 @@
-import { FieldErrors, UseFormHandleSubmit, UseFormRegister } from 'react-hook-form';
+import { UseFormRegister } from 'react-hook-form';
 
-// Form data structure for contact form
 export interface FormData {
   firstName: string;
   lastName: string;
@@ -8,17 +7,6 @@ export interface FormData {
   question: string;
 }
 
-// Props for the contact form section component
-export interface ContactFormSectionProps {
-  register: UseFormRegister<FormData>;
-  errors: FieldErrors<FormData>;
-  handleSubmit: UseFormHandleSubmit<FormData>;
-  isSubmitting: boolean;
-  isValid: boolean;
-  onSubmit: (data: FormData) => Promise<void>;
-}
-
-// Props for the reusable form input component
 export interface FormInputProps {
   label: string;
   name: keyof FormData;
