+++ src/lib/analytics.ts
@@ -1,0 +1,8 @@
+import { AnalyticsBrowser } from '@segment/analytics-next';
+
+export const analytics =
+  typeof window === 'undefined'
+    ? null
+    : AnalyticsBrowser.load({
+        writeKey: process.env.NEXT_PUBLIC_SEGMENT_WRITE_KEY!,
+      });
